refactor: move security email helpers out of users
